Address failing integration tests (#271)

+++ tests/tcp/test_nodes_creation.py
@@ -23,20 +23,10 @@
 import pytest
 
 from ansys.optislang.core import Optislang, node_types
+from ansys.optislang.core.osl_server import OslVersion
 from ansys.optislang.core.tcp.nodes import System
 
 pytestmark = pytest.mark.local_osl
-
-NODE_TYPES_MODULE = dir(node_types)
-NODE_TYPES = [
-    node_type
-    for node_type in NODE_TYPES_MODULE
-    if not (
-        callable(getattr(node_types, node_type))
-        or node_type.startswith("__")
-        or node_type == "annotations"
-    )
-]
 
 
 @pytest.fixture()
@@ -56,15 +46,21 @@ def optislang(scope="function", autouse=False) -> Optislang:
 
 def test_all_nodes_creation(optislang: Optislang):
     """Test creation of all available nodes."""
+    if optislang.osl_version < OslVersion(24, 1, 2, 0):
+        pytest.skip(f"Not compatible with {optislang.osl_version_string}")
+
     rs = optislang.application.project.root_system
     rs.delete_children_nodes()
-    for node_type in NODE_TYPES:
-        print(f"Creating {eval('node_types.' + node_type)}")
-        node = rs.create_node(type_=eval("node_types." + node_type))
-        nodes_in_rs = rs.get_nodes()
-        assert node.uid == nodes_in_rs[0].uid
-        assert node.type == nodes_in_rs[0].type
-        node.delete()
+
+    for node_group in optislang.osl_server.get_available_nodes().values():
+        if node_group == "builtin_nodes" or node_group == "integration_plugins":
+            for node in node_group:
+                print(f"Creating node {node}")
+                node = rs.create_node(type_=node_types.get_node_type_from_str(node))
+                nodes_in_rs = rs.get_nodes()
+                assert node.uid == nodes_in_rs[0].uid
+                assert node.type == nodes_in_rs[0].type
+                node.delete()
 
 
 def test_create_node_in_system(optislang: Optislang):
